update validation to new format

+++ code/validation/validation.R
@@ -1,4 +1,4 @@
-library(covid19SMHvalidation)
+library(SMHvalidation)
 library(gh)
 
 # check if submissions file
@@ -8,8 +8,10 @@ pr_files <- gh::gh(paste0("GET /repos/",
 
 pr_files_name <- purrr::map(pr_files, "filename")
 pr_sub_files <- grep(
-  "data-processed/.*/\\d{4}-\\d{2}-\\d{2}.+-.+(.csv|.zip|.gz)", pr_files_name,
+  "data-processed/.*/\\d{4}-\\d{2}-\\d{2}.+-.+(.csv|.zip|.gz|.pqt|.parquet)", pr_files_name,
   value = TRUE)
+pop_path <- "data-locations/locations.csv"
+js_def_dir <- dir("code/validation/round_metadata/", full.names = TRUE)
 
 # Run validation on file corresponding to the submission file format
 if (length(pr_sub_files) > 0) {  
@@ -22,26 +24,51 @@ if (length(pr_sub_files) > 0) {
     dir.create(paste0(getwd(), "/proj_plot"))
   # run validation and visualization
   test_tot <- lapply(seq_len(length(pr_sub_files_lst)), function(x) {
+    # submission file
     url_link <- URLdecode(pr_sub_files_lst[[x]]$raw_url)
-    if (grepl(".zip$|.gz$", url_link)) {
+    # associated metadata file
+    js_def_file <- grep(stringr::str_extract(basename(url_link), 
+      "\\d{4}-\\d{2}-\\d{2}"), js_def_dir, value = TRUE)
+    if (length(js_def_file) > 1) {
+        if (grepl("sample", basename(url_link))) {
+          js_def_file <- grep("sample", js_def_file, value = TRUE)
+        } else {
+          js_def_file <- grep("sample", js_def_file, value = TRUE, invert = TRUE)
+        }
+    }
+    # Run validation and visualization if associate metadata file found
+    if (length(js_def_file) == 1) {
+      # validation and visualization for ZIP and GZ file format
+      if (grepl(".zip$|.gz$", url_link)) {
       # download file
       download.file(url_link, basename(url_link))
       # generate visualization pdf
       test_viz <- try(generate_validation_plots(path_proj = basename(url_link), lst_gs = lst_gs, 
         save_path = paste0(getwd(), "/proj_plot"), y_sqrt = FALSE, plot_quantiles = c(0.025, 0.975)))
       # run validation
-      test <- capture.output(try(validate_submission(basename(url_link), lst_gs = lst_gs)))
+      test <- capture.output(try(validate_submission(basename(url_link), lst_gs = lst_gs, 
+        pop_path = pop_path, js_def = js_def_file)))
+      }
+      # validation and visualization for CSV and PARQUET file format
+      if (grepl(".csv$|.parquet$", url_link)) {
+        # generate visualization pdf
+        test_viz <- try(generate_validation_plots(path_proj = url_link, lst_gs = lst_gs, 
+          save_path = paste0(getwd(), "/proj_plot"), y_sqrt = FALSE, plot_quantiles = c(0.025, 0.975)))
+        # run validation
+        test <- capture.output(try(validate_submission(url_link, lst_gs = lst_gs, 
+          pop_path = pop_path, js_def = js_def_file)))
+      }
+      # Remove visualization pdf if viz has an error
+      if (class(test_viz) == "try-error") 
+        file.remove(dir(paste0(getwd(), "/proj_plot"), full.names = TRUE))
+    } else {
+      test_viz <- NA
+      test <- capture.output(cat(paste0(
+        "\U000274c Error: The associated metadata file was not found, the validation and visualization was not run. ",
+        "Please verify that the date in the name of the file corresponds to the expected date. Please feel free to",
+        " tag `@LucieContamin` for any question or additional information")))
     }
-    if (grepl(".csv$", url_link)) {
-      # generate visualization pdf
-      test_viz <- try(generate_validation_plots(path_proj = url_link, lst_gs = lst_gs, 
-        save_path = paste0(getwd(), "/proj_plot"), y_sqrt = FALSE, plot_quantiles = c(0.025, 0.975)))
-      # run validation
-      test <- capture.output(try(validate_submission(url_link, lst_gs = lst_gs)))
-    }
-    # Remove visualization pdf if viz has an error
-    if (class(test_viz) == "try-error") 
-      file.remove(dir(paste0(getwd(), "/proj_plot"), full.names = TRUE))
+    
     # list of the viz and validation results
     test_tot <- list(valid = test, viz = test_viz)
     # returns all output
@@ -79,10 +106,11 @@ if (any(!is.na(test_viz))) {
     "[here](https://docs.github.com/en/actions/managing-workflow-runs/downloading-workflow-artifacts)")
 
   if (any(unlist(purrr::map(test_viz, class)) == "try-error")) {
-    message_plot <- paste0(message_plot, "\n\n",
+    message_plot <- capture.output(
+      cat(paste0(message_plot, "\n\n\U000274c Error: ",
       "The visualization encounters an issue and might not be available,",
       " if the validation does not return any error, please feel free to ",
-      "tag `@LucieContamin` for any question.")
+      "tag `@LucieContamin` for any question.")))
   }
 
   gh::gh(paste0("POST /repos/", "midas-network/covid19-scenario-modeling-hub/",
